move JSON formatter into Runtime module
