+++ stdlib/public/libexec/swift-backtrace/Utils.swift
@@ -255,34 +255,3 @@ func formatISO8601(_ time: timespec) -> String {
 
   return isoTime
 }
-
-/// Escape a JSON string
-func escapeJSON(_ s: String) -> String {
-  var result = ""
-  let utf8View = s.utf8
-  var chunk = utf8View.startIndex
-  var pos = chunk
-  let end = utf8View.endIndex
-
-  result.reserveCapacity(utf8View.count)
-
-  while pos != end {
-    let scalar = utf8View[pos]
-    switch scalar {
-      case 0x22, 0x5c, 0x00...0x1f:
-        result += s[chunk..<pos]
-        result += "\\"
-        result += String(Unicode.Scalar(scalar))
-        pos = utf8View.index(after: pos)
-        chunk = pos
-      default:
-        pos = utf8View.index(after: pos)
-    }
-  }
-
-  if chunk != end {
-    result += s[chunk..<pos]
-  }
-
-  return result
-}
